plasma/lss: Fix a race condition between flushbuffer reset and init
A reset was performed by non-atomically setting state variable as there
are no writers resetting it. But, there could be a old delayed reader
which as access to the flushbuffer thinking that it is not closed yet.
When a reset is performed, old accessor thinks that its a valid buffer
and makes allocations into it. But, buffer reuse allocation may not have
happened by that time. This causes multiple accessors to get valid
space reservations with lssOffset = 0.

This commit makes reset operation and buffer reuse assignment atomic by
adding a new reset state. When a buffer is reset, it just sets reset bit
and the buffer is still set as full. Hence, any random accessors won't be
able to make reservation. When reuse assignment is made, it atomically
changes state from reset+full to non-reset+non-full.

Change-Id: Icc907c6c73064567ea24ad7e0dff20acec09882d